requirements.txt
flask
python-dotenv
PyPDF2
langchain_openai
openai
faiss-cpu
langchain_community
langchain
flask-cors